Fix burger (when click on nav-link its closing the mobile menu)
