+++ src/index.js
@@ -2,14 +2,13 @@ import './index.css';
 import handleLike from './modules/handleLike.js';
 import { renderCards, updateLikesCountCallback } from './modules/renderCards.js';
 import fetchData from './modules/fetchShows.js';
-import toggleBurger from './modules/navToggle.js';
+import toggleMenu from './modules/toggleButton.js';
 import { incrementCounter, updateCounter } from './modules/showCounter.js';
-import { adjustCloseBtnPosition } from './modules/adjustCloseBtn';
 
 const container = document.getElementById('card-container');
 let counter = { showsTotal: 0 };
-const navToggle = document.querySelector('.nav-toggle-label');
-const navMenu = document.querySelector('.nav-links');
+const toggleButton = document.querySelector('.nav-toggle-label');
+const navLinks = document.querySelectorAll('.nav-link');
 
 const fetchDataAndRenderCards = async () => {
   const shows = await fetchData();
@@ -28,13 +27,9 @@ document.addEventListener('click', async (event) => {
     await handleLike(itemId, updateLikesCountCallback);
   }
 });
-
-navToggle.addEventListener('click', () => {
-  const visibility = navMenu.getAttribute('data-visible');
-  if (visibility === 'true' && window.innerWidth < 768) {
-    toggleBurger();
-  }
+toggleButton.addEventListener('click', toggleMenu);
+navLinks.forEach((navLink) => {
+  navLink.addEventListener('click', toggleMenu);
 });
 
 fetchDataAndRenderCards();
-adjustCloseBtnPosition();
